+++ server/tls/read-creds-from-cert-files.js
@@ -1,13 +1,6 @@
-const grpc = require('grpc');
 const { readFileSync } = require('fs');
 
-function readCredsFromCertFiles({
-  caPath,
-  keyPath,
-  certPath,
-  serverName,
-  verifyHost,
-}) {
+function readCredsFromCertFiles({ caPath, keyPath, certPath }) {
   if (!keyPath) {
     throw Error('TLS key is not provided');
   }
@@ -15,8 +8,6 @@ function readCredsFromCertFiles({
   if (!certPath) {
     throw Error('TLS certificate is not provided');
   }
-
-  let credentials;
 
   const pk = readFileSync(keyPath);
   const cert = readFileSync(certPath);
@@ -26,27 +17,7 @@ function readCredsFromCertFiles({
     ca = readFileSync(caPath);
   }
 
-  let checkServerIdentity;
-  if (verifyHost) {
-    checkServerIdentity = (receivedName, cert) => {
-      if (receivedName !== serverName) {
-        throw new Error(
-          `Server name verification error: ${serverName} but received hostname ${receivedName}`
-        );
-      }
-    };
-  }
-
-  credentials = grpc.credentials.createSsl(caPath ? ca : undefined, pk, cert, {
-    checkServerIdentity,
-  });
-
-  const options = {};
-  if (serverName) {
-    options['grpc.ssl_target_name_override'] = serverName;
-  }
-
-  return { credentials, options };
+  return { pk, cert, ca };
 }
 
 module.exports = { readCredsFromCertFiles };
