+++ server/utils.js
@@ -2,4 +2,10 @@ const isWriteApiPermitted = () => {
   return ![false, 'false'].includes(process.env.TEMPORAL_PERMIT_WRITE_API);
 };
 
-module.exports = { isWriteApiPermitted };
+function compareCaseInsensitive(a, b) {
+  return typeof a === 'string' && typeof b === 'string'
+    ? a.localeCompare(b, undefined, { sensitivity: 'accent' }) === 0
+    : a === b;
+}
+
+module.exports = { isWriteApiPermitted, compareCaseInsensitive };
